refactor: TextSimilarityUtils가 전역적으로 사용되도록 패키지 변경
