+++ src/main/java/mju/iphak/maru_egg/common/utils/NLP/CustomTFIDFSimilarity.java
@@ -1,4 +1,4 @@
-package mju.iphak.maru_egg.question.utils.NLP;
+package mju.iphak.maru_egg.common.utils.NLP;
 
 import org.apache.lucene.index.FieldInvertState;
 import org.apache.lucene.search.similarities.TFIDFSimilarity;
